Moved run with AD to test, added more dependencies to test env
