+++ src/driver/mpas_ocean.jl
@@ -52,97 +52,9 @@ function ocn_run(config_fp)
     println(clock.currTime)
 end
 
-# Runs forward model with AD, and computes FD derivative approximations for comparison
-function ocn_run_with_ad(config_fp)
-
-    #
-    # Setup for model
-    #
-    
-    #backend = KA.CPU()
-    backend = CUDABackend()
-
-    # Initialize the Model  
-    Setup, Diag, Tend, Prog             = ocn_init(config_fp, backend = backend)
-    clock, simulationAlarm, outputAlarm = ocn_init_alarms(Setup)
-
-    timestep   = KA.zeros(backend, Float64, (1,))
-    d_timestep = KA.zeros(backend, Float64, (1,))
-    sumGPU     = KA.zeros(backend, Float64, (1,))
-    d_sumGPU   = KA.zeros(backend, Float64, (1,))
-
-    sumCPU   = zeros(Float64, (1,))
-    d_sumCPU = zeros(Float64, (1,))
-    @allowscalar timestep[1] = convert(Float64, Dates.value(Second(Setup.timeManager.timeStep)))
-
-    #
-    # Actual Model Run with AD
-    #
-
-    d_Prog = Enzyme.make_zero(Prog)
-    for i = 1:2
-        d_Prog.normalVelocity[i] = KA.zeros(backend, Float64, size(d_Prog.normalVelocity[i]))
-        d_Prog.layerThickness[i] = KA.zeros(backend, Float64, size(d_Prog.layerThickness[i]))
-        d_Prog.ssh[i] = KA.zeros(backend, Float64, size(d_Prog.ssh[i]))
-    end
-
-    d_Diag = Enzyme.make_zero(Diag)
-    d_Diag.layerThicknessEdge = KA.zeros(backend, Float64, size(d_Diag.layerThicknessEdge))
-    d_Diag.thicknessFlux      = KA.zeros(backend, Float64, size(d_Diag.thicknessFlux))
-    d_Diag.velocityDivCell    = KA.zeros(backend, Float64, size(d_Diag.velocityDivCell))
-    d_Diag.relativeVorticity  = KA.zeros(backend, Float64, size(d_Diag.relativeVorticity))
-
-    d_Tend = Enzyme.make_zero(Tend)
-    d_Tend.tendNormalVelocity = KA.zeros(backend, Float64, size(d_Tend.tendNormalVelocity))
-    d_Tend.tendLayerThickness = KA.zeros(backend, Float64, size(d_Tend.tendLayerThickness))
-
-    d_Setup = Enzyme.make_zero(Setup)
-    d_ForwardEuler = Enzyme.make_zero(ForwardEuler)
-    d_clock = Enzyme.make_zero(clock)
-    d_simulationAlarm = Enzyme.make_zero(simulationAlarm)
-    d_outputAlarm = Enzyme.make_zero(outputAlarm)
-    
-    d_sum = autodiff(Enzyme.Reverse,
-             ocn_run_loop,
-             Duplicated(sumCPU, d_sumCPU),
-             Duplicated(sumGPU, d_sumGPU),
-             Duplicated(timestep, d_timestep),
-             Duplicated(Prog, d_Prog),
-             Duplicated(Diag, d_Diag),
-             Duplicated(Tend, d_Tend),
-             Duplicated(Setup, d_Setup),
-             Duplicated(ForwardEuler, d_ForwardEuler),
-             Duplicated(clock, d_clock),
-             Duplicated(simulationAlarm, d_simulationAlarm),
-             Duplicated(outputAlarm, d_outputAlarm),
-             )
-    
-    @show d_Prog.normalVelocity[end][1:10]
-    @show d_Prog.layerThickness[end][1:10]
-
-    #
-    # Writing to outputs
-    #
-    
-    # Only suport i/o at the end of the simulation for now 
-    write_netcdf(Setup, Diag, Prog, d_Prog)
-    
-    backend = get_backend(Tend.tendNormalVelocity)
-    arch = typeof(backend) <: KA.GPU ? "GPU" : "CPU"
-
-    println("Moka.jl ran on $arch")
-    println(clock.currTime)
-end
-
 if abspath(PROGRAM_FILE) == @__FILE__
     if isfile(ARGS[1])
-        if (length(ARGS) > 1 && ARGS[2] == "--with_ad")
-            #include("../../ext/EnzymeExt.jl")
-            #using Enzyme
-            ocn_run_with_ad(ARGS[1])
-        else
-            ocn_run(ARGS[1])
-        end
+        ocn_run(ARGS[1])
     else 
         error("yaml config file invalid")
     end
