+++ test/Project.toml
@@ -7,3 +7,4 @@ KernelAbstractions = "63c18a36-062a-441e-b654-da1e3ab1ce7c"
 MOKA = "4c5738b9-ea62-4abc-b436-bc5c16744d63"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
+YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
